Add files via upload
Rough sketch of wiring in car. Blue squares represent XT60 connectors breaking off of the main power bus. The purpose of these is to allow for us to adjust the length of the wires connecting the subsystems to the bus in the future. If we did not include these, we would have to desolder wires from the main bus. The orange represents this connecting wire, with an XT60 on the bus side and an aviation plug on the subsystem side.